requirements.txt
apturl==0.5.2
asgiref==3.5.0
backports.zoneinfo==0.2.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.1
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
entrypoints==0.3
filelock==3.4.2
gpg===1.13.1-unknown
httplib2==0.14.0
idna==2.8
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Markdown==3.1.1
Nautilus-scripts-manager==2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
packaging==20.3
pbr==5.8.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.1
protobuf==3.6.1
psycopg2==2.9.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.14.1
pyRFC3339==1.1
pysmbc==1.0.15.6
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sqlparse==0.4.2
stevedore==3.5.0
systemd-python==234
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.13.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
xkit==0.0.0
